WIP: protocols/receivers:
It's still not possible to eliminate the listener for UDP just because
we don't store the I/O handler, the thread mask nor the context in the
receiver.

It doesn't make much sense to duplicate all these settings for each and
every receiver for a given bind_conf. There seems to be something wrong
with where the split is performed.

Maybe we should first split the bind_conf to keep only receiver options.
such as unix uid/gid/mode, interface, namespace, bind_thread etc. This
would then become the real bind_conf, and the rest would be the listen_conf.

The storage would look like this :

 proxy --+
         |
         V
     listen_conf  ---------> bind_conf ---> proto
        |||                      ^            |
        VVV                      |            V
     listeners ------------> receiver     addr family

Another option would be that the receivers are actually between the
bind_conf and the listeners. We'd have this :

  proxy ---->>> bind_conf ---> proto
   |||            ||| ^
   VVV            VVV |
  listener <--> receivers